Context switching metrics
As of context switching can have a big impact of application performance
it would be nice to control how often it happens. Context switches are
generally slow operation taking about 100-1000 cycles of CPU and more
over after the switch we can observe CPU cache missing as the memory
cached for previous thread should be invalidated for the current one.
Measuring this value allow us to estimate overheads of using hundreds
of threads in classic blocking application or figure out that something
wrong happens in reactive application.